+++ cpp/tests/centrality/katz_centrality_test.cu
@@ -36,7 +36,7 @@ std::vector<int> getTopKIds(double* p_katz, int count, int k = 10)
 }
 
 template <typename VT, typename ET, typename WT>
-int getMaxDegree(cugraph::experimental::GraphCSR<VT, ET, WT> const& g)
+int getMaxDegree(cugraph::experimental::GraphCSRView<VT, ET, WT> const& g)
 {
   cudaStream_t stream{nullptr};
 
@@ -116,11 +116,10 @@ class Tests_Katz : public ::testing::TestWithParam<Katz_Usecase> {
       << "\n";
     ASSERT_EQ(fclose(fpin), 0);
 
-    CSR_Result<int> result;
-    ConvertCOOtoCSR(&cooColInd[0], &cooRowInd[0], nnz, result);
-
-    cugraph::experimental::GraphCSR<int, int, float> G(
-      result.rowOffsets, result.colIndices, nullptr, m, nnz);
+    cugraph::experimental::GraphCOOView<int, int, float> cooview(
+      &cooColInd[0], &cooRowInd[0], nullptr, m, nnz);
+    auto csr                                               = cugraph::coo_to_csr(cooview);
+    cugraph::experimental::GraphCSRView<int, int, float> G = csr->view();
 
     rmm::device_vector<double> katz_vector(m);
     double* d_katz = thrust::raw_pointer_cast(katz_vector.data());