+++ cpp/tests/sssp/sssp_test.cu
@@ -259,7 +259,7 @@ class Tests_SSSP : public ::testing::TestWithParam<SSSP_Usecase> {
     }
     CSR_Result_Weighted<MaxVType, DistType> result;
     ConvertCOOtoCSR_weighted(&cooRowInd[0], &cooColInd[0], &cooVal[0], num_edges, result);
-    cugraph::experimental::GraphCSR<MaxVType, MaxEType, DistType> G(
+    cugraph::experimental::GraphCSRView<MaxVType, MaxEType, DistType> G(
       result.rowOffsets, result.colIndices, (DistType*)nullptr, result.size, result.nnz);
     if (DoRandomWeights) { G.edge_data = result.edgeWeights; }
     cudaDeviceSynchronize();